+++ cppcache/include/geode/StructSet.hpp
@@ -64,7 +64,7 @@ class _GEODE_EXPORT StructSet : public CqResults {
    * @returns A smart pointer to the item indexed.
    */
   virtual const std::shared_ptr<Serializable> operator[](
-      int32_t index) const = 0;
+      size_t index) const = 0;
 
   /**
    * Get the index number of the specified field name in the StructSet.
